refactor:interface:centralized
