+++ src/components/ListProducts.tsx
@@ -1,5 +1,6 @@
 import { FC } from 'react';
+import { IProductItems } from '../interfaces';
 
-export const ListProducts: FC = () => {
+export const ListProducts: FC<IProductItems> = () => {
   return <p>List {`:=>`} Products</p>;
 };
